Fixes #21
Low Level Keyboard hooks don't include device information. Raw Input Device
listeners do, but they don't allow for blocking events. To combine both, we
run a listener of each. Low level hook events are processed first (unknown
why), so we put those events into a "deviceless_events" queue. When the
corresponding raw input device events comes (identified by order), we fill
the missing device id and pass it along to the generic keyboard library.
Raw input devices are sometimes buffered, but as long as events are served
in order, this should be ok.

Unforutnately the device id is fetched after the window of opportunity
for blocking events, so we can never block events based on it.
On the other hand, we don't need to wait for the device id before deciding
to block or allow, so events are processed more quickly and with less key
delay.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.