Revert pyobjc dependency declaration back to sys_platform
This is related to issues #136 and #143. By declaring the dependency
using "sys_platform" instead of an `if` in setup.py, the egg is built
correctly regardless of the platform used for building the release.

However clients complained of broken installations on Windows and Linux
(#143). To avoid breakage, the `if` was reinstituted.

The problem was that the platform used to build the release has to have
an up-to-date setuptools installed, since this is a somewhat new
feature. Otherwise the dependency is copied verbatim, "; sys_platform==..."
string included. Confusingly, clients facing this garbled dependency
declaration still try to install the package, resulting in errors that
look like the "sys_platform" part was ignored entirely.

This commit reverts back to using "sys_platform", and this developer
promises to be extra careful not to build releases on systems with
outdated setuptools.